+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import QuestionBox from '/Users/jayp/Desktop/application-projects/chatter-box/src/Components/QuestionsSection/QuestionBox.js'
 import SavedResponses from '/Users/jayp/Desktop/application-projects/chatter-box/src/Components/ResponseSection/SavedResponses.js'
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 const { Configuration, OpenAIApi } = require("openai");
 
 
@@ -10,16 +10,34 @@ function App() {
   const [botAnswer, setBotAnswer] = useState('')
   const [responses, setResponses] = useState([])
 
+    useEffect(() => {
+    if(botAnswer && userQuestion) {
+    createResponse()
+    }
+  }, [botAnswer]);
+
+
+
+  const createResponse = () => {
+  let saveResponse = {
+    id: responses.length + 1,
+    question: userQuestion,
+    botAnswer: botAnswer
+  }
+    responses.push(saveResponse)
+    setUserQuestion('')
+    setBotAnswer('')
+   
+  }
+
     const displayResponse = () => {
       let result = responses.map((answer, i) => {
-        return (<SavedResponses key={i} botAnswer={answer.botAnswer}  question={answer.question} />)
+        return (<SavedResponses key={i} botAnswer={answer.botAnswer} setBotAnswer={setBotAnswer}  question={answer.question} />)
       })
-      console.log(result);
       return result
     }
 
     const postQuestion = () => {
-
     const configuration = new Configuration({
       apiKey: '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',
     });
@@ -35,17 +53,14 @@ function App() {
       stop: [" Human:", " AI:"],
     })
     .then((response) => {
-      console.log(response.data.choices[0].text)
-    })
-
+      setBotAnswer(response.data.choices[0].text)
+    }).catch(err => console.log(err))
     }
 
   return (
     <div className="App">
       <QuestionBox 
       postQuestion={postQuestion} 
-      setResponses={setResponses} 
-      responses={responses} 
       botAnswer={botAnswer} 
       userQuestion={userQuestion} 
       setUserQuestion={setUserQuestion} />
