+++ src/Components/ResponseSection/SavedResponses.css
@@ -1,0 +1,5 @@
+.saved-response{
+ margin: 0%;
+ width: 50%
+ 
+}
