+++ src/Components/ResponseSection/SavedResponses.js
@@ -6,8 +6,8 @@ const SavedResponses = ({question, botAnswer}) => {
   return (
    <> 
     <div className='saved-response'>
-     <p>Question:{question}</p>
-     <p>Chatter Bot Answer:{botAnswer}</p>
+     <p>Question: {question}</p>
+     <p>Chatter Bot Answer: {botAnswer}</p>
     </div>
    </>
   )
